Splitted local data source into post, user and comment local data source
