+++ app/src/main/java/co/com/monkeymobile/post_viewer/di/DataSourceModule.kt
@@ -1,7 +1,11 @@
 package co.com.monkeymobile.post_viewer.di
 
-import co.com.monkeymobile.post_viewer.data.source.local.LocalDataSource
-import co.com.monkeymobile.post_viewer.data.source.local.impl.LocalDataSourceImpl
+import co.com.monkeymobile.post_viewer.data.source.local.CommentLocalDataSource
+import co.com.monkeymobile.post_viewer.data.source.local.PostLocalDataSource
+import co.com.monkeymobile.post_viewer.data.source.local.UserLocalDataSource
+import co.com.monkeymobile.post_viewer.data.source.local.impl.CommentLocalDataSourceImpl
+import co.com.monkeymobile.post_viewer.data.source.local.impl.PostLocalDataSourceImpl
+import co.com.monkeymobile.post_viewer.data.source.local.impl.UserLocalDataSourceImpl
 import co.com.monkeymobile.post_viewer.data.source.remote.CommentRemoteDataSource
 import co.com.monkeymobile.post_viewer.data.source.remote.PostRemoteDataSource
 import co.com.monkeymobile.post_viewer.data.source.remote.UserRemoteDataSource
@@ -21,11 +25,17 @@ abstract class DataSourceModule {
     abstract fun providePostRemoteDataSource(source: PostRemoteDataSourceImpl): PostRemoteDataSource
 
     @Binds
-    abstract fun provideLocalDataSource(source: LocalDataSourceImpl): LocalDataSource
+    abstract fun providePostLocalDataSource(source: PostLocalDataSourceImpl): PostLocalDataSource
 
     @Binds
     abstract fun providesUserRemoteDataSource(source: UserRemoteDataSourceImpl): UserRemoteDataSource
 
     @Binds
+    abstract fun provideUserLocalDataSource(source: UserLocalDataSourceImpl): UserLocalDataSource
+
+    @Binds
     abstract fun providesCommentRemoteDataSource(source: CommentRemoteDataSourceImpl): CommentRemoteDataSource
+
+    @Binds
+    abstract fun provideCommentLocalDataSource(source: CommentLocalDataSourceImpl): CommentLocalDataSource
 }
